Refactor krb5asrep format to facilitate OpenCL implementation
